BC-9277 - move DOs to domain/do folder
