+++ apps/server/src/modules/account/domain/do/password-pattern.spec.ts
