+++ apps/server/src/modules/account/domain/index.ts
@@ -1,9 +1,5 @@
-export * from './account';
-export * from './account-save';
-export * from './idm-account';
+export * from './do';
 export * from './interface';
 export * from './mapper';
 export * from './services';
 export * from './type';
-export * from './update-account';
-export * from './update-my-account';
