+++ apps/server/src/modules/account/domain/services/account-idm.service.integration.spec.ts
@@ -6,8 +6,8 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { ConfigModule } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { v1 } from 'uuid';
-import { AccountSave } from '../account-save';
-import { IdmAccount } from '../idm-account';
+import { AccountSave } from '../do/account-save';
+import { IdmAccount } from '../do/idm-account';
 import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../mapper';
 import { AccountServiceIdm } from './account-idm.service';
 import { AbstractAccountService } from './account.service.abstract';
