+++ apps/server/src/modules/account/domain/services/account-idm.service.ts
@@ -6,10 +6,10 @@ import { ConfigService } from '@nestjs/config/dist/config.service';
 import { EntityNotFoundError } from '@shared/common/error';
 import { Counted, EntityId } from '@shared/domain/types';
 import { AccountConfig } from '../../account-config';
-import { Account } from '../account';
-import { AccountSave } from '../account-save';
+import { Account } from '../do/account';
+import { AccountSave } from '../do/account-save';
+import { IdmAccountUpdate } from '../do/idm-account';
 import { FindAccountByDbcUserIdLoggable, GetOptionalIdmAccountLoggable } from '../error';
-import { IdmAccountUpdate } from '../idm-account';
 import { AccountIdmToDoMapper } from '../mapper';
 import { AbstractAccountService } from './account.service.abstract';
