+++ apps/server/src/modules/account/domain/services/account.service.integration.spec.ts
@@ -14,9 +14,9 @@ import { MongoMemoryDatabaseModule } from '@testing/database';
 import { v1 } from 'uuid';
 import { AccountEntity } from '../../repo';
 import { accountFactory } from '../../testing';
-import { Account } from '../account';
-import { AccountSave } from '../account-save';
-import { IdmAccount } from '../idm-account';
+import { Account } from '../do/account';
+import { AccountSave } from '../do/account-save';
+import { IdmAccount } from '../do/idm-account';
 import { ACCOUNT_REPO, AccountRepo } from '../interface';
 import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../mapper';
 import { AccountServiceDb } from './account-db.service';
