+++ apps/server/src/modules/account/domain/services/account.service.spec.ts
@@ -30,11 +30,11 @@ import 'reflect-metadata';
 import { AccountConfig } from '../../account-config';
 import { AccountEntity, AccountEntityToDoMapper } from '../../repo';
 import { accountDoFactory, accountFactory } from '../../testing';
-import { Account } from '../account';
-import { AccountSave } from '../account-save';
+import { Account } from '../do/account';
+import { AccountSave } from '../do/account-save';
+import { UpdateAccount } from '../do/update-account';
 import { IdmCallbackLoggableException } from '../error';
 import { ACCOUNT_REPO, AccountRepo } from '../interface';
-import { UpdateAccount } from '../update-account';
 import { AccountServiceDb } from './account-db.service';
 import { AccountServiceIdm } from './account-idm.service';
 import { AccountService } from './account.service';
